fix(sim): round projectile and particle-weapon math per x87 op
The original runs gameplay at 24-bit x87 precision. These projectile and
fire paths still used double literals, double chains or a lifecycle gate the
game does not have:

- flamers and Bubblegun pass the unwrapped `heading - 1.5707964f` as the
  particle angle, not an atan2 round-trip wrapped into (-pi, pi]
- particle weapons subtract float32 ammo costs (0.1f/0.05f/0.15f), so a
  Blow Torch clip lasts 601 shots and a Bubblegun clip 101
- Shrinkifier `size * 0.65f`, Splitter `angle -/+ 1.0471976f`, Plasma Cannon
  ring radius/angle/offsets round per op
- Ion Rifle chain and Shock Chain angles round each fsub after the wide
  fpatan, from rounded deltas
- rocket acceleration, speed caps, trail decay (0.01f), detonation radius,
  PC24 distance test and 0.1f impulse
- detonation blasts hit every active creature with positive health,
  including Shrinkifier corpses at any lifecycle stage

New opt-in native-oracle tests run `projectile_update` (rocket flight,
detonations, special primary hits), the Shock Chain branch of `bonus_apply`
and the flamer/Bubblegun fire block of `player_update` against both sides.

+++ src/crimson/weapon_runtime/fire_recipes.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import msgspec
 
 from ..effects import ParticleStyleId
+from ..math_parity import f32
 from ..projectiles.types import ProjectileTemplateId, SecondaryProjectileTypeId
 from ..weapons import WeaponId, projectile_type_id_for_weapon_id
 
@@ -80,6 +81,7 @@ type FireMode = PrimaryPelletsMode | SecondaryShotMode | ParticleStreamMode | Mu
 
 class FireRecipe(msgspec.Struct, frozen=True):
     mode: FireMode
+    # `player_update` subtracts a float32 local (`0.1f`/`0.05f`/`0.15f`).
     ammo_cost: float = 1.0
 
 
@@ -119,19 +121,19 @@ FIRE_RECIPE_BY_WEAPON: dict[WeaponId, FireRecipe] = {
     WeaponId.ROCKET_MINIGUN: FireRecipe(mode=SecondaryShotMode(type_id=SecondaryProjectileTypeId.ROCKET_MINIGUN)),
     WeaponId.FLAMETHROWER: FireRecipe(
         mode=ParticleStreamMode(style=None, slow=False),
-        ammo_cost=0.1,
+        ammo_cost=f32(0.1),
     ),
     WeaponId.BLOW_TORCH: FireRecipe(
         mode=ParticleStreamMode(style=ParticleStyleId.BLOW_TORCH, slow=False),
-        ammo_cost=0.05,
+        ammo_cost=f32(0.05),
     ),
     WeaponId.HR_FLAMER: FireRecipe(
         mode=ParticleStreamMode(style=ParticleStyleId.HR_FLAMER, slow=False),
-        ammo_cost=0.1,
+        ammo_cost=f32(0.1),
     ),
     WeaponId.BUBBLEGUN: FireRecipe(
         mode=ParticleStreamMode(style=None, slow=True),
-        ammo_cost=0.15,
+        ammo_cost=f32(0.15),
     ),
     WeaponId.MULTI_PLASMA: FireRecipe(mode=MultiPlasmaFanMode()),
     WeaponId.MINI_ROCKET_SWARMERS: FireRecipe(mode=SwarmerDumpMode()),
